+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file contains "important" commits, small ones will probably not be added.
 
+### 190828 - 1 [1.0.3]
+- Improved terminal
+- Fixed some bugs when closing screens
+
 ### 190827 - 1 [1.0.3]
 - Fixed inline vertical height between the editor lines
 - Added bracket matching and active line colorizer